smart-video-pro/pyproject.toml
[project]
name = "smart-video-pro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "av>=17.0.1",
    "faster-whisper>=1.2.1",
    "google-generativeai>=0.8.6",
    "langcodes>=3.5.1",
    "langdetect>=1.0.9",
    "language-data>=1.4.0",
    "openai-whisper>=20250625",
    "opencv-python-headless>=4.13.0.92",
    "pyqt6>=6.11.0",
    "pytest>=9.0.3",
    "srt>=3.5.3",
    "stable-ts>=2.19.1",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "ultralytics>=8.4.39",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu118"
